display list: a source range names its document by index, not by pointer
`SourceRange` carried `path: Rc<str>` -- a 16-byte fat pointer, per cluster,
to one of a handful of strings. A 2 MB document has 1.5 M clusters, so the
pointers cost far more than the paths they point at, and a cluster is the
structure that dominates per-glyph memory.

The index already existed. `Span::document` is the compiler's index into the
request's document list, and the display list's own `documents` array is that
same list in that same order; `source_of` was resolving that index to a string
once per glyph. A range now carries the index instead:

    SourceRange { document: DocId(u32), start_byte: u32, end_byte: u32 }

12 bytes, and `Copy`. The paths live in one `DocumentPaths` table per display
list, shared by every page behind an `Rc`, and the serialisers resolve through
it -- so the wire format is unchanged: it still carries paths, and the delta
digest still hashes the path string rather than the index.

Two smaller changes fall out and are what take a cluster below 64 bytes:

* `Provenance`'s rare arms are boxed behind thin pointers
  (`Sources(Box<Vec<..>>)`, `Synthetic(Box<String>)`). Both are rare -- one is
  reserved and unused -- and inline they widened every cluster to carry a
  `Vec`/`String` header. `Provenance`: 40 -> 16 bytes.
* `Cluster`'s `text_start_byte`/`text_end_byte` are `u32`. They index the
  *run's* text, which is one word at most.

    SourceRange  32 -> 12      Provenance  40 -> 16
    Cluster      88 -> 56      Item       200 -> 184

56, not the 64 the follow-up was ranked at: the boxed arms let `Provenance`
pack to 16 rather than 24.

## The index has to survive a reordered document list

It is a per-request index, not a global identity, and the incremental cache
outlives a request. A cached block's ranges name documents as the request that
*built* it listed them; a later request may list the same documents in a
different order. The path string was self-describing and an index is not, so
placement now remaps it.

`Relocate { from, to, delta }` carries both halves of that move: ranges in
`from` become ranges in `to` with their offsets shifted by `delta`, and ranges
in any other document are untouched. `AssembledBlock` records the `DocId` it
was assembled against, so `from` is known. This replaces the old "shift the
offsets of ranges whose path string matches" rule with the same rule stated in
indices, plus the remap the string form did not need.

The two places that relocate both take it: `place_item` when a cached block's
line is placed, and `relocate_block` when a cached block is reused. Both used
to return early on `delta == 0`; they now return early only when the
relocation is a no-op in *both* halves, because the document index can move
while the byte offsets do not.

`delta.rs` compares a base page against a new one across two display lists
whose tables may differ in order, so `ranges_match` compares resolved paths
rather than indices -- the same comparison it made before.

\input chains need nothing special: every document of a project is in the
request's list, and a block whose items straddle two documents is never
cached (`block_origin` refuses it), so a cached block has exactly one document
to remap.

## Memory, measured

`memprofile` from #232, against a control at the same base:

    synthetic-500kb   live  367.8 ->  341.8 MiB   VmRSS  651.2 ->  625.9 MiB
    synthetic-2mb     live 1458.4 -> 1352.9 MiB   VmRSS 2569.5 -> 2466.2 MiB

At 2 MB that is -105.5 MiB live and -103.3 MiB resident. It lands on both
copies of every cluster: the walked display list falls 270.0 -> 218.3 MiB and
the render cache 693.6 -> 639.5 MiB. Allocator retention is unmoved
(1111.2 -> 1113.3 MiB).

More than the ~35 MiB this was ranked at on #232, because that estimate
counted the cached copy only.

+++ crates/render-pipeline/tests/amssymb_braces.rs
@@ -32,7 +32,7 @@ fn glyphs(r: &Rendered) -> Vec<(String, u16, f64, f64, u32)> {
         for g in &run.glyphs {
             let c = &run.clusters[g.cluster as usize];
             out.push((
-                run.text[c.text_start_byte..c.text_end_byte].to_string(),
+                run.text[c.text_start_byte as usize..c.text_end_byte as usize].to_string(),
                 g.gid,
                 g.advance_x.to_bp() * 72.27 / 72.0,
                 run.font_size.to_bp() * 72.27 / 72.0,
